+++ app/initializers/torii-provider-arcgis.js
@@ -10,7 +10,7 @@
  * which reopen's the Torii Session, and adds additional methods
  * to it which are useful for ArcGIS Online Sessions
  */
-
+/* eslint no-unused-vars: ["error", { "varsIgnorePattern": "ext" }] */
 import ext from '../ext/torii-provider-arcgis';
 
 export function initialize () {